Integer value check.

integer_check() throws an error if n is not equal to a positive valued integer.

Usage

integer_check(n)

Arguments

n

An integer.

Value

An integer. If no error is thrown during evaluation, then integer_check() returns the n argument.