Shankle Oil Well Lithology

Six petrophysical measures collected by the Kansas Geological Survey by drilling into the B1-C formations of the Shankle well in the Panoma oil and gas field of southwest Kansas.

Usage

well_log

Format

well_log

A data frame with 346 rows and 8 columns:

Facies

1: "Nonmarine Sandstone, 2: "Nonmarine Coarse Siltstone, 3: Nonmarine Fine Siltstone, 4: Marine Siltstone/Shale, 5: Mudstone, 6: Wackestone, 7: Dolomite, 8: Packstone-Grainstone,8: Packstone-Grainstone"

Depth

Well depth in feet

GR

Gamma ray emissions

Rt

Deep resistivity

DeltaPhi

Difference of neutron and density porosity

AvgPhi

Average neutron and density porosity

PE

Photoelectric factor

MnM

Marine/non-marine indicator

Source

Bohling, G. and M. Dubois (2003). An integrated application of neural network and markov chain techniques to the prediction of lithofacies from well logs: Kansas geological survey open-file report 2003-50, 6 p. Group 6.